pyproject.toml
[project]
name = "telegram-ai-agent"
version = "0.1.0"
description = "A production-grade Telegram AI Assitant with RAG, Memory, and MCP tool integrations"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "chromadb>=1.5.0",
    "google-search-results>=2.4.2",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "tiktoken>=0.12.0",
]
